Link to EEE wikipedia page
Might as well call things by their proper names rather than dance around them.

The whole concept of "EEE" is highly dubious.  Of all the examples, perhaps one or two at a stretch involve a company successfully employing it as a strategy to make money.  But they do _unsuccessfully_ try it fairly regularly, and do a lot of damage along the way.  As long as that's true, it does make sense to include it as a motivation here.