Better approach to manage pre-existing Accounts and Node parameters (#333)
The primary goal of these changes are to create a uniform way to specify pre-existing accounts on the nodes referenced in a TestPlan. This may be necessary because:

* a given app at a Node does not have an API by which accounts can be automatically provisioned
* ... or the code to do so hasn't been written, such as when developers only start to use FediTest with their app
* or the developer may want to re-run the same tests with the same accounts that already exist on an instance they are running themselves as "SaaS"
* or the developer may want to re-run the same tests, getting the exact same user names even if they are auto-provisioned, to make debugging easier.

There are several difficulties:

* Each application may need different data to identify an existing account. For example, some may need username and password, others e-mail address and OAuth token etc.
* Communicating what is needed for a given situation / app to the developer is really hard; nobody likes to pour through long documentation (which we probably couldn't keep up-to-date and testing that a given version is up-to-date would be a challenge in itself). This would -- and has for me -- led to significant frustration: "I'm specifying all these things, why isn't it working?"
* Error checking code will be all over the place. Consistent error checking and reporting is not really possible without a lot of repeat code review work.
* Different tests may require a different (and potentially unlimited) number of accounts. The approach should seamlessly scale to a number larger than the pre-existing accounts, by auto-provisioning new accounts or asking the user to do so.

A similar situation applies to non-existing-accounts: similar to accounts, it identifies what could be an account, but is known not to exist. This is required for things such as WebFinger tests against non-existing accounts.

While migrating the ad-hoc mechanisms we had in the code for specifying pre-existing accounts, I also implemented a similar mechanism for Node parameters: what parameters are required or supported on a given Node was not obvious.

Example that illustrates the point (needs better formatting but you get the point):

```
% feditest info --nodedriver MastodonSaasNodeDriver
Node driver name:             MastodonSaasNodeDriver
Description:                  Create a Mastodon Node that already runs as SaaS
Parameters:                   {'app': 'Name of the app', 'app_version': 'Version of the app', 'hostname': 'DNS hostname of where the app is running.', 'verify_api_tls_certificate': 'If set to false, accessing the Mastodon API will be performed without checking TLS certificates.'}
Account fields:               {'userid': 'Mastodon userid for a user (e.g. "joe") (required).', 'email': 'E-mail for the user on an Account. Required for accounts specified with a password (not a token) because logging into Mastodon is by e-mail address.', 'password': 'Password for a user. Required for accounts that also have an e-mail address and no token.', 'oauth_token': 'OAuth token of a user, obtained out of band, for users not specified with e-mail address and password.', 'role': 'A symbolic name for the Account as used by tests (optional).'}
Non-existing Account fields:  {'userid': 'Mastodon userid for a non-existing user (e.g. "joe")  (required).', 'role': 'A symbolic name for the non-existing Account as used by tests (optional).'}
```

In implementation terms, the key changes are:
* a `NodeConfiguration` object which does all the parameter checking, default settings etc as early as possible so TestPlans with invalid parameters fail long before they try to actually run. This means we don't pass a `parameter` dict around any more, but a config object that, when it exists, is known to be internally consistent.
* An `AccountManager` object associated with each Node. Different implementations implement different strategies for how to return requested accounts and/or when and how to provision new ones. While feditest could conceivably be applied to Nodes that run software that do not have the notion of an account, that seems very unusual and isn't worth making things more complicated; `AccountManager` could simply be ignored in that case.
* Information about pre-existing Accounts and Parameters are captured, including their name, description and error checking code in `TestPlanNodeAccountField`, `TestPlanNodeNonExistingAccountField` and `TestPlanNodeParameter`.
* There are several feditest TestPlans that run against the Mastodon API. Not all of them run yet, but some do.